rm asine, mv lgamma
