+++ src/math/mod.rs
@@ -129,7 +129,9 @@ mod jnf;
 mod ldexp;
 mod ldexpf;
 mod lgamma;
+mod lgamma_r;
 mod lgammaf;
+mod lgammaf_r;
 mod log;
 mod log10;
 mod log10f;
@@ -233,9 +235,9 @@ pub use self::jnf::ynf;
 pub use self::ldexp::ldexp;
 pub use self::ldexpf::ldexpf;
 pub use self::lgamma::lgamma;
-pub use self::lgamma::lgamma_r;
+pub use self::lgamma_r::lgamma_r;
 pub use self::lgammaf::lgammaf;
-pub use self::lgammaf::lgammaf_r;
+pub use self::lgammaf_r::lgammaf_r;
 pub use self::log::log;
 pub use self::log10::log10;
 pub use self::log10f::log10f;
