+++ src/math/modf.rs
@@ -2,7 +2,7 @@ pub fn modf(x: f64) -> (f64, f64) {
     let rv2: f64;
     let mut u = x.to_bits();
     let mask: u64;
-    let e = ((u >> 52 & 0x7ff) as isize) - 0x3ff;
+    let e = ((u >> 52 & 0x7ff) as i32) - 0x3ff;
 
     /* no fractional part */
     if e >= 52 {
